+++ index.html
@@ -234,8 +234,7 @@
       <h2 class="data-error__title">Не удалось загрузить данные</h2>
     </section>
   </template>
-  <script src="js/functions.js"></script>
-  <script src="js/main.js"></script>
+  <script type="module" src="js/script.js"></script>
 </body>
 
 </html>
