+++ js/data.js
@@ -1,3 +1,5 @@
+import { getRandomArrayElement, getRandomNumber} from './util';
+
 const PHOTOS_COUNT = 25;
 const MIN_LIKES = 15;
 const MAX_LIKES = 200;
@@ -7,15 +9,6 @@ const DESCRIPTIONS = ['Утро', 'Вечер', 'Котик' ];
 const COMMENTS = ['Всё отлично!', 'В целом всё неплохо. Но не всё.', 'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.', 'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.', 'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.', 'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент.'];
 
 const NAMES = ['Николай', 'Таня', 'Сургей', 'Оля', 'Иван', 'Егор'];
-
-const getRandomNumber = (a, b) => {
-  const lower = Math.ceil(Math.min(Math.abs(a), Math.abs(b)));
-  const upper = Math.floor(Math.max(Math.abs(a), Math.abs(b)));
-  const result = Math.random() * (upper - lower + 1) + lower;
-  return Math.floor(result);
-};
-
-const getRandomArrayElement = (data) => data[getRandomNumber(0, data.length - 1)];
 
 const createMessage = () =>
   Array.from({ length: getRandomNumber(1, 2) }, () =>
@@ -46,6 +39,6 @@ const createFotos = () =>
     { length: PHOTOS_COUNT }, (_, createIndex) => createFoto(createIndex + 1)
   );
 
-console.log(createFotos());
+createFotos();
 
-
+export {createFoto};
