+++ js/script.js
@@ -1,0 +1,6 @@
+import './functions.js';
+import {createFoto} from './data.js';
+
+console.log(createFoto(10));
+
+
