+++ js/util.js
@@ -1,0 +1,10 @@
+const getRandomNumber = (a, b) => {
+  const lower = Math.ceil(Math.min(Math.abs(a), Math.abs(b)));
+  const upper = Math.floor(Math.max(Math.abs(a), Math.abs(b)));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+};
+
+const getRandomArrayElement = (data) => data[getRandomNumber(0, data.length - 1)];
+
+export {getRandomArrayElement, getRandomNumber};
